refactor: Extract logic from `src/data/index.ts` to `loader.ts`
- Moves implementation logic out of the `index.ts` barrel file
- Keeps `index.ts` strictly as a re-export module to improve codebase structure
